+++ src/app/components/App.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
-import { App } from './App';
+import { store } from '../core/store';
+import { App } from '..';
 
 test('renders learn react link', () => {
 	const { getByText } = render(
