+++ src/app/components/App.tsx
@@ -1,7 +1,3 @@
-import React, { Suspense, useEffect, useMemo, useState } from 'react';
-import { Redirect, Route, Switch } from 'react-router';
-import { BrowserRouter } from 'react-router-dom';
-import { Http404 } from './features/errors/404';
 import {
 	AppBar,
 	Box,
@@ -16,17 +12,22 @@ import {
 	Typography,
 	useTheme,
 } from '@material-ui/core';
-import { Link } from 'react-router-dom';
 import {
 	List as ListIcon,
 	MenuOutlined,
 	Timeline as TimelineIcon,
 	ChangeHistory as ChangeHistoryIcon,
 } from '@material-ui/icons';
-import { useDispatch } from 'react-redux';
-import { horseDefsActions } from './features/horse-defs';
-import { Indicator } from './features/indicator';
-import { PedigreeDialog } from './features/pedigree';
+import React, { Suspense, useEffect, useMemo, useState } from 'react';
+import { Redirect, Route, Switch } from 'react-router';
+import { BrowserRouter, Link } from 'react-router-dom';
+
+import { useAppDispatch } from 'src/util';
+
+import { Http404 } from 'src/features/errors/404';
+import { horseDefsActions } from 'src/features/horse-defs';
+import { Indicator } from 'src/features/indicator';
+import { PedigreeDialog } from 'src/features/pedigree';
 
 import './App.css';
 
@@ -52,7 +53,7 @@ const defs: MenuItemDef[] = [
 		label: '牝系図',
 		path: '/mare-line',
 		Page: React.lazy(() =>
-			import('./features/mare-line').then((module) => ({
+			import('src/features/mare-line').then((module) => ({
 				default: module.MareLine,
 			}))
 		),
@@ -62,7 +63,7 @@ const defs: MenuItemDef[] = [
 		label: '家系図(旧)',
 		path: '/family',
 		Page: React.lazy(() =>
-			import('./features/family').then((module) => ({
+			import('src/features/family').then((module) => ({
 				default: module.Family,
 			}))
 		),
@@ -72,7 +73,7 @@ const defs: MenuItemDef[] = [
 		label: '更新履歴',
 		path: '/change-log',
 		Page: React.lazy(() =>
-			import('./features/changelog').then((module) => ({
+			import('src/features/changelog').then((module) => ({
 				default: module.ChangeLog,
 			}))
 		),
@@ -135,7 +136,7 @@ const Footer: React.FC = () => {
 
 export const App: React.FC = () => {
 	const theme = useTheme();
-	const dispatch = useDispatch();
+	const dispatch = useAppDispatch();
 
 	useEffect(() => {
 		dispatch(horseDefsActions.init());
