+++ src/app/index.ts
@@ -1,0 +1,4 @@
+export type { AppAction } from './core/actions';
+export type { AppDispatch, Dependencies, Epic, RootState } from './core/store';
+export { store, registerEpic, useAppDispatch } from './core/store';
+export { App } from './components/App';
