+++ src/features/changelog/components/ChangeLog.tsx
@@ -1,13 +1,13 @@
 import { Paper, createStyles, makeStyles } from '@material-ui/core';
-
 import * as React from 'react';
 import { useEffect } from 'react';
 import Markdown from 'react-markdown';
-import { useDispatch } from 'react-redux';
-
-import 'github-markdown-css';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../../app/store';
+import 'github-markdown-css';
+
+import { RootState } from 'src/app';
+import { useAppDispatch } from 'src/util';
+
 import { changeLogActions } from '..';
 
 const useStyles = makeStyles((theme) =>
@@ -20,7 +20,7 @@ const useStyles = makeStyles((theme) =>
 
 export const ChangeLog: React.FC = () => {
 	const classes = useStyles();
-	const dispatch = useDispatch();
+	const dispatch = useAppDispatch();
 
 	const { article } = useSelector((state: RootState) => state.changeLog);
 	useEffect(() => {
