+++ src/features/changelog/core/epic.ts
@@ -1,10 +1,10 @@
 import { combineEpics } from 'redux-observable';
 import { filter } from 'rxjs/operators';
 
-import type { Epic } from '../../../app/store';
+import type { Epic } from 'src/app';
+import { withIndicator } from 'src/util';
 
 import { changeLogActions } from '..';
-import { withIndicator } from '../../indicator';
 import { fetchChangeLogs } from './logic';
 
 export const fetchDefsEpic: Epic = (action$) =>
