+++ src/features/family/components/FamilyDiagram.tsx
@@ -2,14 +2,15 @@ import * as React from 'react';
 import { useEffect, useMemo, useRef } from 'react';
 import * as d3 from 'd3';
 import * as dagreD3 from 'dagre-d3';
+import { Edge } from 'dagre-d3';
 import * as graphlib from 'graphlib';
 
-import { HorseDef } from '../../horse-defs';
+import { useSelector } from 'react-redux';
+
+import { RootState } from 'src/app';
+import { HorseDef } from 'src/features/horse-defs';
 
 import './FamilyDiagram.css';
-import { Edge } from 'dagre-d3';
-import { useSelector } from 'react-redux';
-import { RootState } from '../../../app/store';
 
 const prepareGraph = (defs: HorseDef[]) => {
 	const nodes = new Map<string, dagreD3.Label>();
