+++ src/index.tsx
@@ -1,10 +1,11 @@
 import { createMuiTheme, MuiThemeProvider } from '@material-ui/core';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import { App } from './App';
-import { store } from './app/store';
 import { Provider } from 'react-redux';
+
+import './index.css';
+
+import { App, store } from './app';
 import * as serviceWorker from './serviceWorker';
 
 const theme = createMuiTheme({});
